Jacques change for property names + Daniels requested change for removing project locator
