+++ src/Cli/dotnet/commands/dotnet-publish/PublishCommandParser.cs
@@ -15,7 +15,6 @@ namespace Microsoft.DotNet.Cli
     internal static class PublishCommandParser
     {
         public static readonly string DocsLink = "https://aka.ms/dotnet-publish";
-        public static readonly string CustomDefaultConfigurationProperty = "PublishRelease";
 
         public static readonly Argument<IEnumerable<string>> SlnOrProjectArgument = new Argument<IEnumerable<string>>(CommonLocalizableStrings.SolutionOrProjectArgumentName)
         {
