+++ src/Common/MSBuildPropertyNames.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace Microsoft.DotNet.Cli
+{
+    static class MSBuildPropertyNames
+    {
+        public static readonly string PUBLISH_RELEASE = "PublishRelease";
+        public static readonly string PACK_RELEASE = "PackRelease";
+        public static readonly string CONFIGURATION = "Configuration";
+        public static readonly string CONFIGURATION_RELEASE_VALUE = "Release";
+        public static readonly string OUTPUT_TYPE = "OutputType";
+        public static readonly string OUTPUT_TYPE_EXECUTABLE = "Exe";
+    }
+}
